Split FixpSession by concern into partials (#139, part 2/2)
Closes #139.

The Gateway had two remaining files over the 500-LOC ceiling from #139's
acceptance criteria:

- src/B3.Exchange.Gateway/FixpSession.cs (881 LOC)
- src/B3.Exchange.Gateway/FixpSession.Inbound.cs (1042 LOC)

Both are mechanical re-organisations of the same partial class —
FixpSession's public surface, fields, locks, and call sites are
bit-identical. No behavioural change; the 376 Gateway.Tests + Core +
Host + SyntheticTrader + ScenarioReplay suites all pass unchanged.

New partial files (each under 500 LOC):

| File                                          | LOC | Holds                                   |
|-----------------------------------------------|-----|-----------------------------------------|
| FixpSession.cs                                | 438 | header, ctor, state, write delegates    |
| FixpSession.Lifecycle.cs                      | 300 | Close/Suspend/Reattach/Reap/Dispose     |
| FixpSession.CancelOnDisconnect.cs             | 181 | CoD props + timer schedule/fire         |
| FixpSession.Inbound.cs                        | 132 | RunReceiveLoopAsync + helpers           |
| FixpSession.Dispatch.cs                       | 337 | DispatchInboundAsync                    |
| FixpSession.Negotiate.cs                      | 193 | Negotiate decode + state-machine apply  |
| FixpSession.Establish.cs                      | 170 | Establish decode + state-machine apply  |
| FixpSession.HeaderValidation.cs               | 204 | throttle + business-header checks       |
| FixpSession.Watchdog.cs                       |  97 | watchdog loop                           |

The thread-safety / lock-hierarchy contract for FixpSession was
already documented in the field declarations at the top of
FixpSession.cs (the _attachLock and _outboundLock XML docs explain
exactly which paths take which lock and what state the recv/send/
watchdog hot paths assume). The split preserves that single source
of truth: the lock fields stay in the main file, every callsite in
the new partials still goes through ApplyTransition / _attachLock /
_outboundLock as before.

Combined with PR #148 (InboundMessageDecoder split), the largest
non-generated .cs file under src/B3.Exchange.Gateway/ is now
EntryPointListener.cs at 438 LOC.
